+++ Source/BugsnagConfiguration.h
@@ -147,9 +147,9 @@ NSArray<BeforeSendSession> *beforeSendSessionBlocks;
     (const BSG_KSCrashReportWriter *_Nonnull writer);
 
 /**
- *  YES if uncaught exceptions should be reported automatically
+ *  YES if uncaught exceptions and other crashes should be reported automatically
  */
-@property BOOL autoNotify;
+@property BOOL autoDetectErrors;
 
 /**
  * Determines whether app sessions should be tracked automatically. By default this value is true.
@@ -254,6 +254,12 @@ __deprecated_msg("This detection option is unreliable and should no longer be us
 
 - (void)addBeforeNotifyHook:(BugsnagBeforeNotifyHook _Nonnull)hook
     __deprecated_msg("Use addBeforeSendBlock: instead.");
+
+/**
+ *  YES if uncaught exceptions should be reported automatically
+ */
+@property BOOL autoNotify __deprecated_msg("Use autoDetectErrors instead");
+
 /**
  *  Hooks for processing raw report data before it is sent to Bugsnag
  */